7-puts_half

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
 
 void puts_half(char *str)
 {
-	int i, length = 0, n;
+	long int i, length = 0, n;
 
 	for (i = 0; *(str + i) != 0; ++i)
 		length = i;
